+++ packages/pwa-kit-react-sdk/src/ssr/universal/components/storefront-preview/index.js
@@ -8,49 +8,8 @@
 import React, {useEffect} from 'react'
 import PropTypes from 'prop-types'
 import {Helmet} from 'react-helmet'
-import {isServer} from '@tanstack/react-query'
+import {detectStorefrontPreview, getClientScript} from './utils'
 
-/** Origins that are allowed to run Storefront Preview. */
-const TRUSTED_ORIGINS = [
-    'https://runtime.commercecloud.com',
-    'https://runtime-admin-staging.mobify-storefront.com',
-    'https://runtime-admin-preview.mobify-storefront.com'
-]
-
-/** Origin used for local Runtime Admin. */
-const DEVELOPMENT_ORIGIN = 'http://localhost:4000'
-
-/** Detects whether the storefront is running in an iframe. */
-const detectInIframe = () => typeof window !== 'undefined' && window.parent !== window.self
-
-/** Gets the parent origin when running in an iframe. */
-const getParentOrigin = () => {
-    if (detectInIframe()) {
-        if (window.location.ancestorOrigins) return window.location.ancestorOrigins[0]
-        // ancestorOrigins does not exist in Firefox, so we use referrer as a fallback
-        if (document.referrer) return new URL(document.referrer).origin
-    }
-}
-
-const isParentOriginTrusted = (parentOrigin) => {
-    return window.location.hostname === 'localhost'
-        ? parentOrigin === DEVELOPMENT_ORIGIN // Development
-        : TRUSTED_ORIGINS.includes(parentOrigin) // Production
-}
-
-/** Detects whether the storefront is running in an iframe as part of Storefront Preview. */
-export const detectStorefrontPreview = () => {
-    const parentOrigin = getParentOrigin()
-    return Boolean(parentOrigin) && isParentOriginTrusted(parentOrigin)
-}
-
-/** Returns the URL to load the Storefront Preview client script from the parent origin. */
-export const getClientScript = () => {
-    const parentOrigin = getParentOrigin() ?? 'https://runtime.commercecloud.com'
-    return parentOrigin === DEVELOPMENT_ORIGIN
-        ? `${parentOrigin}/mobify/bundle/development/static/storefront-preview.js`
-        : `${parentOrigin}/cc/b2c/preview/preview.client.js`
-}
 export const StorefrontPreview = ({
     enabled = typeof window !== 'undefined' ? Boolean(window.STOREFRONT_PREVIEW?.enabled) : false,
     getToken
@@ -63,12 +22,13 @@ export const StorefrontPreview = ({
                 getToken
             }
         }
-    }, [enabled])
+    }, [enabled, getToken])
     if (!enabled) {
         return null
     }
     isHostTrusted = detectStorefrontPreview()
-
+    console.log('enabled', enabled)
+    console.log('isHostTrusted', isHostTrusted)
     return (
         <>
             <Helmet>
